requirements.txt
flask
flask-restful
pandas
matplotlib
xlrd
gunicorn
json
hashlib
bcrypt
mysql-connector-python
